Minor fixes for cast warnings

+++ vcpp/unistd.h
@@ -366,14 +366,14 @@ typedef int Atom;
 
 inline
 int gethostname(char *name, size_t len)
-{   DWORD bufsize = len;
+{   DWORD bufsize = (DWORD) len;
     BOOL ok = GetComputerNameA(name,&bufsize);
     return ok? 0:-1;
 }
 
 inline
 int getlogin_r(char *buf, size_t len)
-{   DWORD bufsize = len;
+{   DWORD bufsize = (DWORD) len;
     BOOL ok = GetUserNameA(buf,&bufsize);
     return ok? 0:-1;
 }
